Expand on error handling and refactor config parsing
Make UserInteractError related error types more descriptive so that
the caller can actually determine the details about what exactly went
wrong. Before, UserInteractError was just a nondescriptive type that
could only signal the presence of an error. Now, UserInteractError can
indicate what went wrong more precisely.

Refactor the config parser so that the keys used to access data from the
config are dynamic instead of static. Before, the caller would access
the parsed config data output by the parser through a struct that
contained each possible config field hardcoded in as a field. Now, the
parsed config data is stored as a map with an enum as a key, instead of
a hardcoded struct.
This makes it easier to add new config fields, and makes it possible to
automate the error handling of unpresent config fields inside the config
parsing logic instead of in the caller.
This refactoring was mostly done to make detailed error handling as
detailed above less complicated.
Although this refactoring is composed of smaller changes that each do
their own thing, this is the major overarching theme behind the refactoring.