Change config parsing so that it is more safe
Make the config parsing fail fast if the value of a key is different
than expected.
Before, an error would only occur when the key was
actually queried by the calling code, which meant that if the type of
the value of a key was different from expected, no error would occur
unless that key was actually queried.
But now, the error will at the time the config file is parsed,
regardless of whether the key is queried or not.