Improve handling of when config file is inaccessible
Make it so that it is valid for the config file to not exist.
Add a flag to the config file info struct that denotes if it is
valid for the config file to not exist / if the config file must exist.
If this flag is false (it is ok for config to not exist), then the
program also checks the command line arguments to see if a command line
flag denoting the same thing is set. If either of these two things are
set, then the config file must exist. But if not, then it will be valid
for the config file to not exist. If it doesn't exist, then the program
will behave as if an empty config file was provided.
This logic only applies to the presence of the config file in the
filesystem. If the config file is inaccessible due to some other reason
like inadequate permissions, the program will still raise an error
regardless.

Also add a dedicated error variant that represents inaccessibility of
the config file, either due to it not existing (provided config must
exist), permission errors, etc.