Greatly simplify config handling
Switch to the serde crate for config handling. Get rid of almost all old
config handling code and replace it with much simpler counterparts using
the serde crate. This switch makes config handling easier to work with
and much less complicated.
The current config handling logic also reduces flexibility somewhat
because it does not allow for the keys or config file to be
semi-optional. Well, technically it does, but it does not allow for
detailed error handling in the case where a non-existent key was
queried. To get this functionality, you would have to make a custom type
implementing Deserialize. Maybe do this in the future.