Fix issues with project deployment system
The problem with the old setup was that the install script should be run
as root, but the build script has to run as the user that calls the
install script (most likely not root).
To fix this, make it so that there is a dedicated deploy script, which
both builds and installs the files, instead of doing the building in the
install script.