Create a new subproject called fullsystemimage
fullsystemimage is a program that manages what I call full system
images, which are images of an entire operating system.
fullsystemimage allows one to backup a whole system by creating a full
system image of it, and restore it on another computer through the
image. The restored system will have exactly the same files and
directories as the old system. As such, this can be used to transplant a
system from one computer to another.

Create a new rust package to house the code, and start coding it.
Also create a script file, but this probably won't be used.